backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
feedparser==6.0.10
google-analytics-data==0.18.2
google-auth==2.25.2
gunicorn==21.2.0
pytest==7.4.3
